+++ app/src/main/java/com/example/playlistmaker/domain/db/FavoritesInteractor.kt
@@ -4,8 +4,8 @@ import com.example.playlistmaker.domain.models.Track
 import kotlinx.coroutines.flow.Flow
 
 interface FavoritesInteractor {
-    fun isFavorite(trackId: Int): Flow<Boolean>
+    fun isFavorite(trackId: Long): Flow<Boolean>
     fun getFavorites(): Flow<List<Track>>
     suspend fun add(track: Track)
-    suspend fun remove(trackId: Int)
+    suspend fun remove(trackId: Long)
 }
