+++ app/src/main/java/com/example/playlistmaker/domain/db/FavoritesRepository.kt
@@ -4,8 +4,8 @@ import com.example.playlistmaker.domain.models.Track
 import kotlinx.coroutines.flow.Flow
 
 interface FavoritesRepository {
-    fun isFavorite(trackId: Int): Flow<Boolean>
+    fun isFavorite(trackId: Long): Flow<Boolean>
     fun getFavorites(): Flow<List<Track>>
     suspend fun addToFavorites(track: Track)
-    suspend fun removeFromFavorites(trackId: Int)
+    suspend fun removeFromFavorites(trackId: Long)
 }
